+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ require 'bundler/setup'
 require 'dotenv/load'
 require 'ruby_bots'
 require 'simplecov'
+require 'pry'
 
 SimpleCov.start